requirements.txt
anyio==3.5.0
APScheduler==3.9.1
asgiref==3.5.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
dataclasses==0.6
decorator==5.1.1
evidently==0.1.46.dev0
fastapi==0.74.1
fonttools==4.30.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.1
ipython==7.32.0
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.3.2
llvmlite==0.38.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
numba==0.55.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
parso==0.8.3
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.6.0
prompt-toolkit==3.0.28
pydantic==1.9.0
Pygments==2.11.2
pyodbc==4.0.32
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
shap==0.40.0
six==1.16.0
slicer==0.0.7
sniffio==1.2.0
starlette==0.17.1
statsmodels==0.13.2
tenacity==8.0.1
threadpoolctl==3.1.0
tqdm==4.63.0
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.17.5
wcwidth==0.2.5
wincertstore==0.2
zipp==3.7.0
